feat(geral): added Controllers, new Services, New Types, new Enums, and new ClientEnviroment File
created a ClientEnviroment Folder, and a "ClientEnviroment.ts" file that contains informations about
the client (the navigator of user), like the language of the navigator. Created "Controllers"
folder, that have a new Controller, "RegisterController" that will call and methods of
Register.Service, perform some actions, and then call the new Service, "ErrorMesseger.Service", the
job of this Service is to receive messages, and based on the browser language, handle and respond to
the application, added new rules to "AccountRules.ts", created "Interfaces" folder with 2 new
interfaces, "Login.Interface" and "Register.Interface", the Register Controller implements the
"Register.Interface". Added new DTO, "ServerResponseDto.ts", this DTO will receive the response from
the API